Add Eigencraft redstone optimization.
This patch implements theosib's redstone algorithms to completely
overhaul the way redstone works. It should be non-invasive as most of
the code is in the helper class RedstoneWireTurbo.

To enable this optimization, set eigen-redstone to true in sponge global
config. Make sure to disable panda redstone if you want to use this
optimization instead.

More information on how the optimization works can be found below
https://bugs.mojang.com/browse/MC-11193
https://bugs.mojang.com/browse/MC-81098