Cleanup handling for unregistered tile entities
Previously, we were automatically unboxing the Integer returned from
TILE_ENTITY_IDS.get(). If the Integer was null, an 'impossible'
NullPointerException would be thrown, which would appear to come from a
line where an NPE should never happen.

To fix this, we simply remove the special-casing for
SpongeTileEntityType. Any TileEntityType that has not properly been
registered will have a 'fake' id used for timings purposes.