[Mostly] Update Multi-World for 1.13.
With 1.13, Mojang enforces the contract of real DimensionType instances (instead of the magic numbers of the past). Sponge multi-world had
to be rebuilt around this concept all while taking into consideration the desire to have, for example, multiple OVERWORLDs/etc. The engine
has no concept of "GlobalDimensionType"s so I added this concept to support re-using dimensions. The bonus here is our dimension configs and the API
are left none-the-wiser that under the hood each world is it's own DimensionType.

There are still several TODOs, namely re-writing MixinSaveHandler, finishing MixinWorldInfo, and restoring the Copy/Rename/Delete actions.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>