Lord of the Blocks: The Pistons of the Marshes
Migrated what was existing on the EnhancedPhaseContext to
MultiBlockCaptureSupplier. All of the code is optional only
to enable transaction tracking through the specific PhaseStates
that will utilize it. Unique and specific transactions that are
defined showing their tracked original snapshot allows us to
rollback correctly.

Rollbacks are still somewhat finicky, but it is very close to
being handled correctly.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>