Improve overall block property CPU performance.
* Cache block property hashcode.
Block properties are immutable so there is no need to recompute the same hashcode repeatedly.
Instead, we can compute the hashcode once and cache the result.

* Overwrite equals to avoid checking values.
Block properties are unique per block type so we can simply do an instance check.

These changes should reduce the amount of CPU consumption and improve overall performance.

This commit is based on Paper's commit here
https://github.com/PaperMC/Paper/commit/2f8bdbfde84cea2e385aad576ca466c3516db0be