Lord of the Blocks: The Taming of Eventgol
Most of this is cleanup post merging and clarifying some things
about using packet frames instead of constant frame manipulation in
phase states. There is still the question of verifying BlockEventQueue
phase being necessary. As it stands, it serves that the block event
being queued does lack enough information, but that information should
be passed by the current PhaseContext, and not needed to be stored
on a whole new PhaseContext and PhaseState.

More to come. Fixed a few bugs with invalid transactions still being
applied.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>