Lord of the Blocks: The Logic Gate is Closed
Part of this is functional to the respect of being
able to cancel pistons fully, without issue of the client
being aware of the simulation. In near perfect replication
of cancelling the block changes from pistons performing
movements, without the added issue of requiring to listen to
ChangeBlockEvent.Pre with piston context keys. Only issue
is some odd notification propogation not taking place
after the fact.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>