Lord of the Blocks: The Neighbor on the Block
Refactored the special handling of per tile and neighbor transactions as
proper transactions. The side effect of these transactions are that they
need to be post applied without bulk capturing, or risking the duplication
and deeply nested transaction capturing that could prove to cause stack
overflows all over again. The flip of the "immediate" versus "normal" post
process is that the "normal" capturing will take place for all states that
are not considered requiring "special" processing (like BlockEvent, or
TileEntityTick).

So far, it's been verifiable that the entire transaction process is
proceeding semi-normally, with the exception of "instant flip" pistons
being awkwardly broken where the piston attempting to retract a block that
it potentially pushed will result in an "air" block being placed, and the
original block that was intended to be retracted with the sticky piston
will quite literally disappear. The reasoning for this I am assuming is
due to the nature of various getBlockState checks for Pistons as blocks
and TileEntities that will reset the blocks to "air" if they're not done
in the proper order.

The good thing is that cancellations still work without having the
required ChangeBlockEvent.Pre event listener with context key checks.
Another positive change is that the previous injections for Pistons
specifically to perform neighbor notifications are now inlined with the
appropriate IPhaseStates to be thrown without special casing.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>