Lord of the Blocks: The Forbidden Tile
Fixed 0-tick pistons disappearing. There is still an issue somewhere down
the line of processing, whether it's with scheduled block events, or
neighbor notifications that ends up causing complex 0-tick pistons to get
the wrong block state. This is causing not so much of a duplication, but
irregular and non-deterministic block changes that otherwise would not
occur in vanilla. For single reactionary 0-tick piston contraptions that
do not passively trigger other 0-tick pistons, this branch is working. For
complex multi-reactionary 0-tick pistons that chain off one another, this
is still failing and potentially disappearing blocks or moving blocks when
they're not supposed to be moved.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>