Fix chunk packet data length inconsistencies caused by async lighting.
This should fix the issue that the calculated chunk data length is no longer valid when the chunk data extraction begins. This is caused by concurrent modifications caused by the async lighting. Copying the array prevents modifications once we start calculating the data length. The light data of the block storages in the
array can still be modified concurrently while calculating and extracting the data, but this doesn't affect the length of the data so this shouldn't cause problems. Adding/removing block storages from the array is causing inconsistencies.