Lord of the Blocks: The Stairs of Block Transaction
Long story short, in resolving furnaces, 0-tick pistons broke again...
What this needs is a full rewrite of how BlockTransactions are processed,
and queued, and how the Proxy can respect what should exist where and when,
while the queued changes will take place or not.

Biggest change so far is that instead of creating the SpongeblockSnapshot within
the TrackingUtil or PhaseTracker for bulk capturing, the snasphot is created
at the MixinChunk#setBlockState, so that the snapshot can have the included
TileEntity information, and immediate access to whatever data is needed, including
but not withstanding, notifier and owner information. Likewise, this revamps
the creation of BlockTransaction.ChangeBlock to occur as soon as a snapshot can
be created, and only if the IPhaseState#capturesTileChanges returns true. This
refactor does not function fully yet, but once it has been theory proven with the
Proxy to work in insertion order, processing in insertion order, and cancelling
in reverse order, the overall result should be for a good testing ground to start
verifying that Furances will work correctly in switching block states, and
Pistons will switch correctly when in 0-tick pulsed mode.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>