Lord of the Blocks: The States of Master Unwind
Got most all exceptions out of the way with the new handling and automatic
re-entry of multi-captures with UnwindingPhaseState. It itself now fully
re-enables a new supplier if needed, and will only process those sort of
suppliers if the unwinding state is not requiring those changes. Note that
while processing a MultiBlockCaptureSupplier with BlockTransactions
requires that the UnwindingPhaseState not actually capture anything, and
just let the game "play through" as a depth first iteration processing,
while capturing only block changes during the various transaction steps.

There is potential for us to be able to perform the necessary operations,
except that the biggest stopper is actually performing the
NeighborNotification logic, which previoulsy would only call to the
PhaseTracker, but now it seems that it will need to apply to the
BlockTransaction itself instead.

Last closing thoughts before the grand rewrite of Proxy access: Unwinding
is now cleaner than before in that we are taking advantage of various
state flags necessary for unwinding to properly handle itself when
entering from a hybridized IPhaseState that could handle neighbor
notifications, but not tile entities.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>