Default implement a majority of World specific API.
This default implementation process will stay much as it is right now
where the API implementation methods are strictly kept in mixins that are
for implementing the API, while hooks and injections and implementation
specific interface implementations will be made available through a
secondary (normal) mixin.

MixinWorld_API vs MixinWorld.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>