Lord of the Blocks: The Pressing of the Sticky Piston
So, big change... Nuked the ordered transaction lists out of orbit and
added a pseudo linked list implementation to MultiBlockCaptureSupplier to
process the ordered transactions by treating them as nodes. This
implementation should provide us with better handling of intermediary
block states and tile entity states between transactions as one issue that
arose was in the middle of processing block changes, previously queued
changes during captures would contaminate the block state physics and tile
entity handling during each transactions process.

An example of this occuring was with pistons calling to notify neighboring
blocks, and one of the sanity checks Mojang added to the piston block is
that it will get the current block state from the world at the time of
processing. Previous to this commit, the last (most recent) block state
that was applied at the target position of the neighbor notification,
would be returned, not the block state that existed at the time. To
aleviate this contamination, when a new BlockTransaction is being logged,
the "previous states" of either a TileEntity or IBlockState will be
submitted for caching to the earliest possible BlockTransaction that does
not match the same position (and therefor being overwritten), and therefor
we can push these states to the proxy just before the handling of the
transaction.

There is still some neighbor notification handling that needs to be fixed,
and I can verify that furnaces (tile entity replacement with itself after
changing it's block state) works as expected. Due to the nature of
redstone, there is some complicated logic to resolve that is causing
redstone wire not to update after passive power is returned. There's also
a potential stack overflow being caused by this issue.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>