Lord of the Blocks: The Passing of the Block Notification
Wrote out the use of LinkedHashMap in favor of a minimized doubly linked
node implementation in MultiBlockCapture and BlockTransaction. The two are
paired, and being able to remove transactions is now possible (something
that is required for compatibility between this system and the List
implementation system for Forge's captures). Almost done, there's some
awkward push/pop block state issues that end up resulting with invalid or
unexpected states to exist that end up with some pistons to remove
themselves.

Once the block push and popping is fixed with proxy and transactions, this
should resolve all sorts of issues.

Also, added an implementation of ChangeBlockEvent.Grow that should work
for interacted or forced growth. Blocks that also tick and are instances
of IGrowable and the blocks they place are considered growth.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>