Lord of the Blocks: The Muster of Redstone
Added some sanity checks by pushing existing states and tiles to
transactions own mappings. I need to figure out a better way than
propogating it's own original states to itself on construction. An option
would be adding yet another method to the transaction itself to have a say
whether the block state or tile is to be added or not.

Changed up thread field referencese by adding a new static final instance
of the PhsaeTracker to facilitate some enhanced debugging that I can do
with validating once mctester is merged in (validate between what the
client sees versus what the server sees).

Fixed some neighbor things, but there are some last transactional changes
that are not clearing themselves or they are not being sent to clients.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>