Fix post processing accidentally capturing using the enhanced tracking methods.
Typically, we don't want special processing states to be tracked with enhanced
tracking since they are being applied. However, for states that do not have
enhanced tracking, the unwinding of those states needs to be processed correctly
with BlockTransactions.

The other issue thst is removed tile entities were not being properly removed
and occasionally due to the enhanced tracking, multiple removals would be
requested, leading to potentially storing a null entry in the Chunk.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>