Merge SpongeForge changes into Common's processRightClickBlock changes.
Since SpongeForge adds minimal hooks to keep in line with Forge's hooks,
the changes made here are adding those same hooks to SpongeImplHooks and
performing the same sort of logical code that would otherwise have been
contained in SpongeForge. This brings in line bug fixes between SpongeCommon
and SpongeForge that were missing due to this confusing fact of "keeping
the two overwrites" in sync.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>