Fix two bugs: Placing tile entities when SF is on the client, and breaking containers missing their inventories.
For the first: Verified that Sponge was not allowing the client to create
the necessary TileEntities on MixinChunk when client side (the creation was
wrapped in a server side check).

The second issue: TileEntities such aas Chests need to be able to access
their TileEntities when their containing blocks are broken, in which case
they attempt to access their inventory. This is where the issue came around
with the TileEntity already being marked as "removed" prior to actual processing
such that the block.breakBlock method would get a null TileEntity, instead
of the actual TileEntity. Moving the TileEntity queued removal to after the
block.breakBlock call in BlockTransaction allows us to attempt to fetch that
TileEntity that was queued as removed. Furthermore, since the TileEntity was
queued for Removal, the issue then arises that the TileEntity needs to be
"pulled from" the removal queue to actually access when world.getTileEntity
is called during block.breakBlock. To do this, I've enabled another hook for
UnwindingState to override specifically to allow us to pull that TileEntity
from the removal queue. Of course, this doesn't actually pull the entity
out of the queue, it will still remove the TileEntity, but it still resides
in the proxy.

Fixes SpongePowered/SpongeForge#2701.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>