Ensure we are exiting the popping of PhaseContexts when we are in an error state. This should fix the cascading failures that lead to spam with recursive NeighborNotificationState entries.
The root of the issue may have already been resolved with the commit for
SpongePowered/SpongeForge/commit/fdf88ef, however, in the case of
cascading failures, this issue specifically targeted the issue where we
had excessive PhaseContexts being closed out that were no longer in the
stack, leading to NPE's and also erroneous closures of other
PhaseContexts.

This ensures that NeighborNotificationContexts are entered, regardless
whether the max processing depth has been reached, ensuring that the
processing depth will end up leading to a safe exit without popping the
parent PhaseContexts. Also, this ensures that when the stack is popped
for being a runaway PhaseState, we do not continue processing, or
calling a duplicate `this.stack.pop()` after we've already popped.

Fixes the errors causing an ultimate crash.
Fixes SpongePowered/SpongeForge#2706

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>