Fix re-entrant single block captures on UnwindingState causing neighboring blocks not being updated on clients.
This specifically fixes an issue where a block break being processed could lead to
multiple block breaks being processed singularly leading to a re-assignment of the
single snapshot field on PhaseContext. Specifically this rewrites the UnwidingContext
to dynamically push these snapshots to a stack when re-entrance is attempted, and thusly
pushing to said stack, and popping on retrieval.

Fixes SpongePowered/SpongeForge#2700.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>