Fix multiple requests at removing the same TileEntity while breaking blocks. Fixes SpongePowered/SpongeForge#2725.
To give a better explanation: Since we process the block break during
block state sets, we have the potential case where the mod blocks are
replicating vanilla's design for Furnaces where the TileEntity is being
removed during the block break, and then re-set back onto the position
just after the block state set. The issue here was that because an
initial tile entity removal request would be made, the ChangeBlock
transaction would already have a set TileEntity to be removed, and
therefor, a new RemoveTileEntity transaction would be made and processed
later. Because of this, the supplier now checks if the queued removal is
re-duplicating the same TileEntity removal request (again, because
Block#breakBlock already has a world.removeTileEntity call, and so does
Chunk#setBlockState).

This hopefully should resolve many machine related issues where
TileEntities getting cleared out on block state changes.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>