Fix hit vector calculation point
Previous, BlockPos#add(double x, double y, double z) wss called.
However, this truncates the result to a normal integer BlockPos, causing
the computed hitvec to always be the same as the block position.

To fix this, we first convert the BlockPos to a Vector3d before adding
the interaction point, ensuring that we compute the proper result