Fixed a massive issue with recursive block processing for Unwinding.
Specifically, there was an issue that the UnwindingPhaseState would push
new MultiBlockCaptureSuppliers for it's own depth first processing, but
as a side effect, the first "original" supplier would not have it's
Proxy pruned and cleared. The issue would compound on itself after
multiple placements due to the nature of how the Proxy access is
designed to short circuit block state getters during processing of
individual complex transactions. The larger issue would end up leading
to over-extended-uptime bugs that could affect TileEntity processing, as
well as interactions at a later point.

This has been minimally tested to verify that the previous interactions
of furnaces, pistons, etc. still function, but other more obscure bugs
that may have been fixed may need to be re-tested.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>