Revert "Fix block drops requesting a potentially removed TileEntity during player interactions. Fixes SpongePowered/SpongeForge#2709."
Tile entities are not supposed to be available during harvesting. See BlockContainer (the base class for most vanilla blocks with tile entities) as an example. When removing a
block, Block.breakBlock() is called before Block.harvestBlock(). In BlockContainer, the tile entity is deleted in breakBlock(). Mods should obtain the old tile entity through the
parameter passed to harvestBlock().

This reverts commit 5f19dd24c9969574b90d36caa1552015a383d94a.