Fix custom item entities not being spawned without any event listeners.
Specifically put: Forge adds custom item entities by listening to
EntityJoinWorldEvent always, and will process the custom item entity as
a replacement of the original EntityItem. The problem here is thaat
since Sponge overtakes and replaces the hooks with a ShouldFire, well,
the custom item entities don't get spawned for custom items when there
are NO event listeners. This fixes it by adding the embedded hook for
the case where Shouldfire is not used.

Also fixes a compile error introduced in the previous commit.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>