Fix the schematic translator failing to pick up mod BlockEntities.
Due to the nature that mod registered TileEntities are still funky in
1.12, we have to rely on the basis that the id could potentially not
show up in Sponge's related systems, so we have to attempt to find it
through the TileEntity registry itself, and then find the TileEntityType
appropriately. It's possible there's more of an issue with that
registration process, but for the sake of getting schematics to function
with mod entities as well as vanilla, this will suffice.

Fixes SpongePowered/SpongeForge#2761.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>