Fix loading v1 Schematics and translating them for v2 format.
This fixes loading v1 Schematics that were being saved erroneously
by older SpongeCommon implementations. The format specifically stated
using "Id" for the TileEntity type id, but the implementation was
allowing vanilla's "id" as the string key to be used. Because of the
case sensitivity nature of the format, the v1 schematics were being
translated but dropping the TileEntity information, because that
type id would be missing.

The fix for this is just an involvement with the upgrader from v1 to
v2, where it checks for the "id" usage and moves that data to "Id".

v2 storage and reading follows the format correctly.

Also fixes a hiddden bug with TileEntityType registration for mods
where Forge started requiring registry names for TileEntity halfway
through 1.12.2's development cycle, so mods were setting their id's
correctly with their mod id already included in the ID, but now,
Forge ensures that the mod id is prepended, so Sponge does not have
to attempt to prepend the id.

Fixes #2268.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>