Bump the max depth processing by default instead, when we reach neighbor notifications.
After doing some analysis, there's more to neighbor notifications that
can end up causing a printout as a one off, not just from redstone, but
from mods as well. The issue with max-block-processing was that we needed
to be able to stop processing more neighbor notifications to avoid a
StackOverflowError, and the only way to do that is putting a cap. That
cap unfortunately can be reached quite easily with some mods and vanilla
performing excessive re-entrant notifications in a small area, resulting
in spam for hundreds of lines, without really being an exception.

So, I've bumped the max depth to 1000 by default, and put a change for
default processing configs where they were previously 100 to 1000. The
answer about StackOverflowError will be when 1000 is too large (shouldn't
be, some research has said somewhere around 8000 stack depths are allowed,
we wouldn't be reaching that unless someone else adds several chains).

Fixes #2252.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>