Cleanup SpongeEntityType usage. Fix damaging custom entities.
The majority of this cluster of changes is dealing with custom entities
overriding EntityLivingBase#damageEntity and expecting their method to
be called during EntityLivingBase#attackEntityFrom. Since Sponge goes as
far as rewriting that whole logic to provide functional equivalents of
"here's the damage function, you can replace it, or nuke it, or simulate
it." we also need to allow for custom mod entities to simply flow
through the same hook, but allow Sponge's modifications to still work.

To do this, I've extended the SpongeEntityType to hold more specific
information on the entity type class it chaperones, with another bit of
handy reflection, like we do for Block and TileEntity tracking states.

So far, this fixes two bugs:
- Fixes SpongePowered/SpongeForge#2546
- Fixes SpongePowered/SpongeForge#2719

In reviewing the changes, I stumbled on another bug that has gone
unnoticed since
https://github.com/SpongePowered/SpongeCommon/commit/deefb43fd0115aea915fbaeaafbb93c77ddc5c4b#diff-3563ea4883521a85084697204735b589R384
was committed. When getting the ItemStack to apply damage for helmet
shielding, we were using the main hand item, not the helmet slot item.

Afterwards, removing the anonymous implementation of EntityType in
SpongeEntityType and turning that into a proper SpongeEntityType
subclass, this cleaned up a bit of the field addition creeping going on
in MixinEntity and MixinEntityLivingBase. In doing so, I took the
opportunity to clean up the shadowed methods and fields for legibility.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>