Clean up runaway explosion spam. Some mods just can't keep control of themselves...
That was sarcasm.

Much like the previous issue where neighbor notifications spam from
redstone was just crashing the server, this issue itself is centered
around the potential case where a block on explosion ends up causing
more explosions. The problem with this is that the phase entry wasn't
aware that the explosions were all different. Only by checking the phase
context for each explosion being actually different is it going to know
whether it's not a runaway.

If the case stood where a *duplicate* explosion at the same position and
world was occurring, then we'd be in a potential runaway candidacy.

As an example, the printout here shows that we have multiple explosions
at different positions triggering eachother, but none of them
duplicating.
https://gist.github.com/gabizou/b886028f334c0c9be5255cecb8ab2121

This fixes SpongePowered/SpongeForge#2750.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>