Sanity check all accesses to getting a World out of a BlockSnapshot.
An unknown bug caused by some "legacy" code in a certain plugin exposed
the possibility of the PhaseTracker capturing block changes from a world
that is about to be unloaded, and naturally, once the world is unloaded,
can no longer be found, leading to a very nice exception that should
never happen. The world unloading was fixed in a prior commit, but the
sanity checks in the Transaction process should be done.

This changes all accesses to a WorldServer from a SpongeBlockSnapshot to
sanity checked and cased in some places to emit a warning that will be a
single line. The advantage of this is that BlockTransactions now
validate that the world they are working on is actually valid, and no
longer prone to throwing an exception that would otherwise fill the
console with a massive message. Because the only remaining information a
transaction has about it's target world is the UUID, the messages would
only be able to provide said UUID, not the world name or dimension.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>