Don't ignore classes that end up with errors when looking up methods.
Some mod classes end up retaining a method that has arguments or return
types that are classes defined only in the client. Because the
Reflection API ends up finding those methods, it will bomb out and fail
to find the class, ending in the silent "do nothing" catch blocks. Now,
we will just catch and try to do a simple equality check on this class
versus the Block class, and if it is available, it will just enable the
three booleans.

This overall fixes the compatibility issue with BetterWithMods.

Fixes SpongePowered/SpongeForge#2770.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>