Clean up world/server tick states.
For some reason I didn't think of it before to merge the classes for pre
and post tick states on world and server ticking. Obviously, server
ticking requires that we do not perform bulk captures, due to the nature
of potentially working on several worlds/unloaded-worlds at once. This
should fix SpongePowered/SpongeForge#2608.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>