Split up Inventory implementation mixins
This is a larger refactor of simply splitting where the inventory
mixins are creating/shadowing/soft-overloading fields by themselves,
whether they are adding a new field or were potentially overwriting
a field added by a previous mixin was not known. Now, the fields added
by InventoryTraitContainerAdapterMixinProducerAbstractConsumerImplementationProviderBridgeMixin.....
No, ok, the fields are only added by InventoryTraitContainerAdapterMixin
or by implementation specific mixins (such as SpongeForge) that will
call to lazily generate their Lens, Fabric, and SlotProvider. This has
been tested with the implementation of ContainerMixin calling methods
to perform SlotTransaction creation, and all nullability of the Lenses
has been rendered obsolete.

There is more to clean up with this implementation as the adapters are
still unused in or somewhat confusing with how they're supposed to be
used versus how they were originally designed. Zidane and Faithcaio will
be taking charge to revamping the implementation of Inventories for
cleaner implementation and handling of all IInventory instances.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>