Don't fail inventory actions if Entity's modid is unknown.
This fixes edge cases where mods due really weird things with Entities
(such as using ASM to add child classes) and Sponge doesn't pick up on
it. Of course a plugin won't be able to rely on grabbing the
PluginContainer of the inventory but at least they can act on the
inventory itself.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>