Add warnings when the level_sponge.dat file isn't saved properly
We are not yet sure as to what the reason is for file saves not working. Much of the hypothesis is centred around the file saving failing at a critical moment, but there is another potential reason, the NBT compound we're saving is empty. I am unsure as to why this would be, so to diagnose this, this commit adds two warnings:

* If the NBT compound we're trying to save is empty, don't bother saving
* If the file is zero sized but the NBT compound was not empty, don't replace the current level file

If it is the former, it explains the reason behind the old file also being empty.

See #2328

Also, if there is no level_sponge.dat, don't remove the level_sponge.dat_old on saving