Re-add Bridge method to SaveHandlerBridge and cleanup.
SaveHandlerBridge had the method bridge$loadSpongeDatData which on the
surface was only used by SaveHandlerMixin, and was rightfully removed as
an interface method, however, on updating SpongeVanilla, the vanilla
mixin was also needing to use that method to load the Forge dimension
data (since Sponge follows Forge's dimension mapping scheme by contract
for compatibility). So, re-adding the method allowed for a SpongeVanilla
update. However, additionally, on inspection of SaveHandlerMixin, I've
migrated some of the string constants to the Constants file, to further
reduce redundancy and allow for compiler inlining. This is not much of a
change for SpongeForge as the visibility of the bridge method wasn't
used in SpongeForge.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>