Re fix #2325. Fixes Vanishing and Griefing states being cached in NBT
The original issue was that Players and other entities would potentially
re-save a false "true" state of being vanished, and on respawning, the
player entity would be marked as vanished still, even though a plugin
could have already set them to not being vanished. The issue was
workaround fixed by always saving the vanishing state on EntityMixin,
however, the actual issue isn't from always saving, it's from the fact
that Entities have their own cached NbtTagCompound that already had the
vanishing state being set, and that compound wasn't re-created every
time the entity is being saved.

This properly fixes and now allows for the sponge compound to be
properly pruned of useless data that defaults to an empty compound, and
properly fixes the cached state being stored in the compound, especially
when vanish is being un-set.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>