A few fixes: More optional saving data, ChunkBridge method.
Two more fixes:
- ChunkBridge unimplemented method fix. Migrated a few private methods
that were added to ChunkMixin_API that should not exist there, moved to
ChunkMixin and properly prefix named the methods. Fixes #2349.
- EntityPlayerMP not optionally removing health scaling when the scale
was reset. Eliminated more optional data and ensured that the cached
value on the Sponge data was properly pruned when necessary. Performed
otherwise some normal house cleaning on the mixin class, eliminating
warnings and removing a redundant class SkinUtil.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>