Fix inventory reverting on relog
When a plugin accessed/modified a offline User in such a way that it got marked as dirty
Sponge would save the user when the player logs in.
When saving/invalidating the User data during login the old inventories
(SpongeUserInventory/SpongeUserInventoryEnderChest) were not removed
causing the inventory to revert to the state before the first login.

fixes #2346