Revert changes to Configurate translators.
There have been numerous issues with this, with issues surrounding deserialization and data loss.

Note that #2353 will replace this - and will need to integrate these changes again. It is not yet complete,
as I do not know how long it will take to get it ready I have elected to back this out, the previous bugs that
this originally was meant to fix are the lesser of two evils at this moment.