Enhance vanishing capabilities. Fixes #2366.
Specifically allow for the EntitySelectors.NOT_SPECTATING to be
modified by Sponge to allow detection of entities that are vanished as
"non spectators". In the case of chests, when a player is opening a
chest, the TileEntity will perform a bounding box lookup for any
available players that are not spectators, and then play their opening
sound and animation. Likewise, for EntityAI, there's a lot of targeting
going on with the selectors in newer versions that would almost
undoubtedly still target vanished entities in some corner cases.

The side effects now are that some interactions may no longer function
for vanished players, such as dispensing armor to a vanished player, or
other typed predicated filters.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>