Release Sponge 7.1.7
This marks the latest release of SpongeAPI's implementation
targeting SpongeForge and SpongeVanilla.

Due to the extended duration since the latest release,
a full change log will not be provided.

The majority of the implementation has had a partial
code audit along with all Mixin cleanup. More can
be read about these changes in #2292 and #2310 as
well as the wiki on our naming policies. Along with
a bundle of various bug fixes, the implementation is
considered "recommended" compared to the previous
release.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>