We'll be storing if world data is mod created.
In 1.12, we based this on the DimensionType. This was not correct then
as nothing stopped a plugin from creating a world through the API that
used a mod DimensionType. This mistake is compounded by the fact that
Sponge-created worlds will be MCP DimensionTypes in 1.14. Easiest fix is
to store this state on our own.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>