Have Vanilla/Mod GameRule changes affect all worlds.
Sponge's Multi-World concept unfortunately breaks the assumption that
mods have that all adjustments to GameRules will reflect on all worlds.
Since we give each world it's own SaveHandler, we need to restore that
assumption but at the same time we also need to ensure that plugins that
want to do per-world GameRules still work. This commit is a happy
compromise.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>