Fix interacts with fake players.
Mods, such as Thaumcraft, may use a fake player to interact with a block during an entity tick.
To handle this scenario, if a fake player is detected during an interaction event, the fake player
will be added as context so the proper ticking source can be used as root cause.