Revert TileEntity capture changes (for now).
As we've attempted to fix issues with mods that interact with TEs, we have come up against numerous issues which have not really fixed many, if any of the original issues, and have also caused duplications in other mods and vanilla. They are therefore being reverted on the main branch for now.

The following commits have been reverted, which means the original issues that they were meant to repair should also be considred open.

* 6b28e62f549f88a0b3b51b8abba91dc90e1eec5a
* 3480230127f27b3df4c2a3e3b5643a14cb25cb4b
* 85f1a678271c803448517d4b31a1348e72df0d76
* 6ae1dab9039b4e72edd52743c962bf42350219a2

Fixes #2847