Move a majority of non-compiling complicated to an invalid folder
Since the majority of the project is beginning to get cleaned up,
we have an over abundance of packages/features that are not required
to get the game launching, and some legacy code still exists overall
that is no longer needed. This speeds up the process of finding the
invalid files that need major work to get compiling again, and moving
them to a temporary home where they can be reviewed when those features
are being worked on.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>